requirements.txt
pylint==2.15.5
pytest==7.2.0
requests==2.28.1
vcrpy==4.2.1
